requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
google-generativeai==0.3.2